Train test models instead of reading from incompatible model files
